requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.11.0
anyio==4.12.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
boto3==1.42.26
botocore==1.42.26
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
conda-pack==0.8.1
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
DataProperty==1.1.0
datasets==4.4.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
einops==0.8.1
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.127.0
ffmpy==1.0.0
filelock==3.20.1
fire==0.7.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
gast==0.7.0
gguf==0.17.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core
idna==3.11
interegular==0.3.3
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.3.1
llvmlite==0.46.0
lm-format-enforcer==0.10.12
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
math-verify==0.9.0
matplotlib==3.10.8
mbstrdecoder==1.1.4
mdurl==0.1.2
mistral_common==1.8.4
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multipledispatch==1.0.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
ollama==0.6.1
omegaconf
openai==2.15.0
orjson==3.11.5
oss2==2.19.1
ouro-cache-fix==0.1.0
outlines==0.0.34
outlines_core==0.2.10
packaging==25.0
pandas==2.3.3
parse==1.20.2
partial-json-parser==0.2.1.1.post6
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge_score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.4.3
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlitedict==2.1.0
sse-starlette==3.0.4
starlette==0.50.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tempdir==0.7.1
torchdata
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.3.0
tomlkit==0.13.3
tqdm==4.67.1
transformers==4.56.0
tree-sitter==0.25.2
tree-sitter-python==0.25.0
trl==0.9.4
typeguard==4.4.4
typepy==1.3.4
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.3
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
wget==3.2
word2number==1.1
xformers
xgrammar==0.1.21
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zstandard==0.25.0
